:new: Add main module
