+++ .eslintrc
@@ -1,3 +1,6 @@
 {
-  "extends": "steelbrain"
+  "extends": "steelbrain",
+  "rules": {
+    "import/prefer-default-export": 0
+  }
 }
